fix: use double for progressPercentage to match test expectations
- Changed progressPercentage from BigDecimal to double in GoalResponse
- BigDecimal.stripTrailingZeros() was producing scientific notation (5E+1)
  and plain '0' instead of expected '50.0' and '0.0'
- double naturally serializes as 0.0, 50.0, 65.5 matching test script

+++ src/main/java/com/aaryav/finance/dto/response/GoalResponse.java
@@ -15,6 +15,6 @@ public class GoalResponse {
     private LocalDate targetDate;
     private LocalDate startDate;
     private BigDecimal currentProgress;
-    private BigDecimal progressPercentage;
+    private double progressPercentage;
     private BigDecimal remainingAmount;
 }
